Refactor redirectr storage implementations
* Move storage implementations to not automatically required ruby files
* Move Redirectr::ReferrerToken to separate file (but still required automatically)
